requirements.txt
aioredis==1.3.1
aiosmtplib==1.1.6
anyio==3.2.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.5.2
fastapi==0.65.2
fastapi-mail==0.3.7
h11==0.12.0
hiredis==2.0.0
httpcore==0.13.6
httptools==0.2.0
httpx==0.18.2
idna==2.10
iniconfig==1.1.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==20.9
passlib==1.7.4
peewee==3.14.4
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
python-decouple==3.4
python-dotenv==0.18.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
websockets==9.1
Werkzeug==2.0.1
